Moving known error checks one call level in
