+++ lib/agent0/agent0/hyperdrive/crash_report/__init__.py
@@ -5,3 +5,4 @@ from .crash_report import (
     log_hyperdrive_crash_report,
     setup_hyperdrive_crash_report_logging,
 )
+from .known_error_checks import check_for_invalid_balance, check_for_slippage
